Move towards consistent scala & C type definitions and more complete posix.
This pull request requires the prior acceptance of PRs #1243, & #1252.

  * Alphabetized in many places to ease checking things that should match
    actually do.  This is software engineering, not code churn.

  * Tuned types.scala and types.h for _LP64 & __ILP32__ data models. That
    is, 32 bit integers on both 32 & 64 bit hardware with longs and pointer
    sizes matching  architecture bit count.  Added code to top of each
    to indicate the other should match or, at least, be consistent.

    Someday one file will be generated from the other but this is the
    best that can be done now to ensure the two points of truth match.

  * Modified stat.scala and unistd.scala to declare and use the
    foundational types in types.scala. The type declarations
    in stat.scala and unistd.scala are a move towards better compliance.

  * Populated stat.scala and unistd.scala with more, but not all,
    posix declarations. Added supporting code to stat.c and unistd.c.

    Removed many simple scalanative @name declarations and corresponding
    C code. This to make it easier to tell when the scalanative C layer is
    actually doing a transform and reduces both source and generated code
    sizes.  Other files could probably benefit from this change.

  * Added the AT_* types to fcntl.scala and fcntl.c. These are n
    needed by newly declared unistd.scala linkat().

  * A changing to unistd.scala read() to the posix definition revealed
    an issue on 64 bit systems in javalib FileInputStream.scala.
    On those systems unistd read() can return 64 bits while java
    can only handle 32 bits. Type checking revealed the mismatch, yeah!
    The truncation (.toInt) now in that file is safe. The length argument
    passed to unistd.read is an Int argument to the enclosing function
    so unistd.read will return a long, but the value in that long
    will always fit into an int.

64/32 bit issues:

      types.scala definitions are tuned to work on 64 bit systems
      and 32 bit X86 systems.

      In particular, I have _NOT_ tested on ARM. There is robustness code
      in resources types.h to FAIL the compilation on untested 32 bit
      systems.

Documentation:

      No user facing documentation changes.

Testing:

	"sbt test-all" passes on Linux 64 bit systems & 32 bit X86.
	I was not able to test on OSX or other operating systems or
	hardware.

+++ javalib/src/main/scala/java/io/FileInputStream.scala
@@ -52,7 +52,8 @@ class FileInputStream(fd: FileDescriptor) extends InputStream {
     // we use the runtime knowledge of the array layout to avoid
     // intermediate buffer, and write straight into the array memory
     val buf       = buffer.asInstanceOf[runtime.ByteArray].at(offset)
-    val readCount = unistd.read(fd.fd, buf, count)
+    // toInt never causes data loss on 64 bit systems because count is an Int.
+    val readCount = unistd.read(fd.fd, buf, count).toInt
 
     if (readCount == 0) {
       // end of file
